# api.yaml
components:
  schemas:
...
    CurrentUserResponseModel:
      properties:
        birthday:
          format: date
          type: string
        email:
          format: email
          maxLength: 64
          type: string
        first_name:
          maxLength: 64
          type: string
        is_admin:
          type: boolean
        last_name:
          maxLength: 64
          type: string
        other_name:
          maxLength: 64
          type: string
        phone:
          type: string
      required:
      - birthday
      - email
      - first_name
      - is_admin
      - last_name
      - other_name
      - phone
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          $ref: '#/components/schemas/ValidationError'
      type: object
    LoginModel:
      properties:
        password:
          maxLength: 128
          type: string
        username:
          maxLength: 64
          type: string
      required:
      - password
      - username
      type: object
...
    Registration:
      properties:
        additional_info:
          maxLength: 256
          type: string
        birthday:
          format: date
          type: string
        city:
          type: integer
        email:
          format: email
          maxLength: 64
          type: string
        first_name:
          maxLength: 64
          type: string
        last_name:
          maxLength: 64
          type: string
        other_name:
          maxLength: 64
          type: string
        password:
          maxLength: 128
          type: string
        phone:
          type: string
        username:
          maxLength: 64
          type: string
      required:
      - email
      - first_name
      - last_name
      - password
      - username
      type: object
...
    ValidationError:
      properties:
        loc:
          type: string
        msg:
          type: string
        type:
          type: string
      required:
      - loc
      - msg
      - type
      type: object
info:
  title: My App
  version: 1.0.0
openapi: 3.0.3
paths:
...
  /login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginModel'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponseModel'
          description: OK
          headers:
            Set-Cookie:
              schema:
                description: Set both access_token_cookie and refresh_token_cookie
                type: string
        '400':
          description: Bad request
        '401':
          description: Wrong credentials
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      tags:
      - auth
  /refresh:
    post:
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponseModel'
          description: OK
          headers:
            Set-Cookie:
              schema:
                description: Set both access_token_cookie and refresh_token_cookie
                example: access_token_cookie=eyJ0eXAi...; Path=/; HttpOnly
                type: string
        '401':
          description: Password changed since token was issues
      tags:
      - auth
  /register:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Registration'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponseModel'
          description: OK
        '400':
          description: Bad request or weak password
        '409':
          description: Username already in use
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      tags:
      - auth
  /logout:
    post:
      responses:
        '200':
          description: OK
          headers:
            Set-Cookie:
              schema:
                description: Set both access_token_cookie and refresh_token_cookie
                example: access_token_cookie=deleted; path=/; expires=Thu, 01 Jan
                  1970 00:00:00 GMT
                type: string
        '401':
          description: Password changed since token was issues
      tags:
      - auth
...